Refactor app to be domain based for actions and reducers
