+++ src/RootReducer.js
@@ -1,5 +1,5 @@
 import {combineReducers} from 'redux';
-import user from './FetchUserReducer';
+import user from './User/Reducers/FetchUserReducer';
 
 export default combineReducers({
     user
