+++ src/User/Actions/FetchUserActions.js
@@ -1,4 +1,4 @@
-import {UserAPI} from '../User/User.api'
+import {UserAPI} from '../API/User.api'
 
 export const FETCH_USER_LOADING = 'FETCH_USER_LOADING';
 export const FETCH_USER_SUCCESS = 'FETCH_USER_SUCCESS';
