+++ src/store.js
@@ -1,6 +1,6 @@
 import {createStore, applyMiddleware} from 'redux';
 import thunk from 'redux-thunk';
-import RootReducer from './Reducers/RootReducer'
+import RootReducer from './RootReducer'
 
 export default function store() {
     return createStore(
